minor changes and refactors; cleanup;
