+++ server/constants/index.ts
@@ -1,8 +1,10 @@
 // * API specific constants
-export const GRANT_TYPE = 'grant_type';
-export const PASSWORD_TYPE = 'password_type';
 export const CLIENT_ID_KEY = 'client_id';
 export const CLIENT_SECRET_KEY = 'client_secret';
+export const GRANT_TYPE = 'grant_type';
+export const PASSWORD_TYPE = 'password_type';
 
 // Content-Type header constant
 export const CONTENT_TYPE = 'application/x-www-form-urlencoded';
+export const CONTENT_TYPE_CHARSET =
+  'application/x-www-form-urlencoded; charset=utf-8';
