+++ server/utils/auth.utils.ts
@@ -1,0 +1,5 @@
+import { CLIENT_ID, CLIENT_SECRET } from '../config';
+
+export const base64Credentials = Buffer.from(
+  `${CLIENT_ID}:${CLIENT_SECRET}`,
+).toString('base64');
